+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.1.0] - 2025-07-25
+
 ### Added
 
 - Transform deprecated `z.object().merge()` to `z.object().extend()`
@@ -123,6 +125,7 @@ For more information about Zod v4 changes, see the [official migration guide](ht
 
 <!-- Links -->
 
+[1.1.0]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.0.2...1.1.0
 [1.0.2]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/nicoespeon/zod-v3-to-v4/compare/84de37227de3b85c13458b6637b57b7cc95a799b...1.0.0